queue: More compact syntax
This patch allows to use a new syntax more compact and break
the current syntax. This new syntax is more similar than the nftables
syntax that we use usually. We can use range like we have used in
other case in nftables. Here, we have some examples:

Before, If we want to declare a queue, we have used a syntax like this:

 nft add rule test input queue num 1 total 3 options bypass,fanout

If we want to use the queue number 1 and the two next (total 3),
we use a range in the new syntax, for example:

 nft add rule test input queue num 1-3 bypass fanout

Also if we want to use only one queue, the new rules are like:

 nft add rule test input queue num 1	# queue 1

or

 nft add rule test input queue		# queue 0

And if we want to add a specific flags we only need to put
what flags we want to use:

 nft add rule test input queue bypass

we don't need to use options and the comma for indicating the
flags.

Signed-off-by: Alvaro Neira Ayuso <alvaroneay@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>