build: remove unnecessary libintl.h check
configure.ac checks for libintl.h which is not used and
may cause unnecessary trouble with e.g. embedded toolchains.

The only reference to libintl.h can be found in parser.c
as generated by bison however an include does not happen
as parser.h defines YYENABLE_NLS to be 0.

Signed-off-by: Steven Barth <cyrus@openwrt.org>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>